pk값 타입 수정 Long -> Integer

+++ src/main/java/kr/co/shoulf/web/entity/Position.java
@@ -6,8 +6,6 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Entity
 @Table(name = "position")
 @Data
@@ -17,7 +15,7 @@ import java.util.List;
 public class Position {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long positionNo;
+    private Integer positionNo;
     @Column(length = 100, nullable = false)
     private String bigName;
 }
